Added scrollPhysics to data_table_2
In this update, the scrollPhysics property has been added to the SingleChildScrollView in data_table_2.dart file. This new feature enhances the usability of the table view, by allowing data rows to be scrollable vertically even if the table view is not fully displayed. This addition can be particularly beneficial when used in combination with the RefreshIndicator.